// src/packages/components/Charts/Maps/MapTmap/type.ts
export enum coverEnum {
  Marker = "Marker",
  Polygon = "Polygon",
  Circle = "Circle",
  Polyline = "Polyline",
}
export enum ThemeEnum {
  NORMAL = '',
  BLACK = 'black',
  INDIGO = 'indigo'
}
export enum contorlEnum {
  Zoom = 'Zoom',
  Scale = 'Scale',
  Copyright = 'Copyright',
  OverviewMap = 'OverviewMap',
  mapType  = 'MapType'
}
export enum mapStyleKeyCustomOptionsLabelEnum {
  SLDT_LNGLAT_SHADOW = '矢量底图经纬度投影',
  SLDT_MOKATUO_SHADOW = '矢量底图球面墨卡托投影',
  SLZJ_LNGLAT_SHADOW = '矢量注记经纬度投影',
  SLZJ_MOKATUO_SHADOW = '矢量注记墨卡托投影',
  YXDT_LNGLAT_SHADOW = '影像底图经纬度投影',
  YXDT_MOKATUO_SHADOW = '影像底图墨卡托投影',
  YXZJ_LNGLAT_SHADOW = '影像注记经纬度投影',
  YXZJ_MOKATUO_SHADOW = '影像注记墨卡托投影',
  DXYX_LNGLAT_SHADOW = '地形晕渲经纬度投影',
  DXYX_MOKATUO_SHADOW = '地形晕渲墨卡托投影',
  DXZJ_LNGLAT_SHADOW = '地形注记经纬度投影',
  DXZJ_MOKATUO_SHADOW = '地形注记墨卡托投影',
  QQJJ_LNGLAT_SHADOW = '全球境界经纬度投影',
  QQJJ_MOKATUO_SHADOW = '全球境界墨卡托投影',
  SLYWZJ_LNGLAT_SHADOW = '矢量英文注记经纬度投影',
  SLYWZJ_MOKATUO_SHADOW = '矢量英文注记墨卡托投影',
  YXYWZJ_LNGLAT_SHADOW = '影像英文注记经纬度投影',
  YXYWZJ_MOKATUO_SHADOW = '影像英文注记墨卡托投影',
}
export enum mapStyleKeyCustomOptionsUrlEnum {
  SLDT_LNGLAT_SHADOW = 'vec_c',
  SLDT_MOKATUO_SHADOW = 'vec_w',
  SLZJ_LNGLAT_SHADOW = 'cva_c',
  SLZJ_MOKATUO_SHADOW = 'cva_w',
  YXDT_LNGLAT_SHADOW = 'img_c',
  YXDT_MOKATUO_SHADOW = 'img_w',
  YXZJ_LNGLAT_SHADOW = 'cia_c',
  YXZJ_MOKATUO_SHADOW = 'cia_w',
  DXYX_LNGLAT_SHADOW = 'ter_c',
  DXYX_MOKATUO_SHADOW = 'ter_w',
  DXZJ_LNGLAT_SHADOW = 'cta_c',
  DXZJ_MOKATUO_SHADOW = 'cta_w',
  QQJJ_LNGLAT_SHADOW = 'ibo_c',
  QQJJ_MOKATUO_SHADOW = 'ibo_w',
  SLYWZJ_LNGLAT_SHADOW = 'eva_c',
  SLYWZJ_MOKATUO_SHADOW = 'eva_w',
  YXYWZJ_LNGLAT_SHADOW = 'eia_c',
  YXYWZJ_MOKATUO_SHADOW = 'eia_w',
}